tres

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -11,7 +11,7 @@
 
 int **alloc_grid(int width, int height)
 {
-	int **a;
+	long int **a;
 	int i, j;
 
 	if (width < 1 || height < 1)
